requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Database — REMOVED (MySQL no longer used)
# mysql-connector-python>=9.1.0

# Data Fetching
yfinance>=0.2.50
aiohttp>=3.9.0

# Data Analysis
pandas>=2.2.3
numpy>=2.1.0
# pandas-ta has compatibility issues with Python 3.14, using ta-lib alternative or manual calculations
# pandas-ta>=0.3.14b0

# Optional: ta-lib (if preferred over pandas-ta)
# TA-Lib==0.4.28

# Utilities
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0

# AI Stock Research Engine Dependencies
# NLP and Sentiment Analysis
transformers>=4.35.0
torch>=2.1.0
vaderSentiment>=3.3.2
sentence-transformers>=2.2.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# FREE Web Research - RSS Feed Parser
feedparser>=6.0.10

# Email
# SMTP is built-in, no additional package needed
